% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_make.R
\name{validate_make}
\alias{validate_make}
\title{Generate a validate program}
\usage{
validate_make(path = file.path(validateMake::get_desktop(), "TestCore"))
}
\arguments{
\item{path}{TestCore path}
}
\description{
Generate a validate program
}

